Change IPD device file to /dev/input-device
Previously, the IPD device was hardcoded to the /dev/ttyUSB0 location.
This often works as intended, but not always. We cannot guarantee that
/dev/ttyUSB0 is the input device, because it might happen that another
device claims it before the input device can do it itself.

This is a problem when we plug the 4G USB modem and the IPD in at
the same time. The 4G USB modem would sometimes claim /dev/ttyUSB0,
which would cause this launch file to fail because the IPD would be
located in a different device file.

This commit fixes this issue by defaulting the device file to
/dev/input-device. This device does not exist by default. It is created
by a custom udev rule that detects when the input device is connected.
When it is, /dev/input-device is created as a symbolic link to the
actual device file.

The udev rule:
KERNEL=="ttyUSB[0-9]*", SUBSYSTEM=="tty", ATTRS{idVendor}=="1a86", ATTRS{idProduct}=="7523", SYMLINK+="input-device"

If you put this in /etc/udev/rules.d/83-input-device.rules, then the
device file will automatically be created when you plug the input device
in.